+++ examples/vue-query/src/gen/hooks/useFindPetsByStatus.ts
@@ -1,79 +1,60 @@
-import client from '@kubb/plugin-client/client'
-import type { FindPetsByStatusQueryResponse, FindPetsByStatusQueryParams, FindPetsByStatus400 } from '../models/FindPetsByStatus'
-import type { RequestConfig } from '@kubb/plugin-client/client'
-import type { QueryKey, QueryObserverOptions, UseQueryReturnType } from '@tanstack/vue-query'
-import type { MaybeRef } from 'vue'
-import { queryOptions, useQuery } from '@tanstack/vue-query'
-import { unref } from 'vue'
+import client from "@kubb/plugin-client/client";
+import type { FindPetsByStatusQueryResponse, FindPetsByStatusQueryParams, FindPetsByStatus400 } from "../models/FindPetsByStatus";
+import type { RequestConfig } from "@kubb/plugin-client/client";
+import type { QueryKey, QueryObserverOptions, UseQueryReturnType } from "@tanstack/vue-query";
+import type { MaybeRef } from "vue";
+import { queryOptions, useQuery } from "@tanstack/vue-query";
+import { unref } from "vue";
 
-export const findPetsByStatusQueryKey = (params?: MaybeRef<FindPetsByStatusQueryParams>) => [{ url: '/pet/findByStatus' }, ...(params ? [params] : [])] as const
+ export const findPetsByStatusQueryKey = (params?: MaybeRef<FindPetsByStatusQueryParams>) => [{ url: "/pet/findByStatus" }, ...(params ? [params] : [])] as const;
 
-export type FindPetsByStatusQueryKey = ReturnType<typeof findPetsByStatusQueryKey>
+ export type FindPetsByStatusQueryKey = ReturnType<typeof findPetsByStatusQueryKey>;
 
-/**
+ /**
  * @description Multiple status values can be provided with comma separated strings
  * @summary Finds Pets by status
  * {@link /pet/findByStatus}
  */
-async function findPetsByStatus(
-  {
-    params,
-  }: {
-    params?: FindPetsByStatusQueryParams
-  },
-  config: Partial<RequestConfig> = {},
-) {
-  const res = await client<FindPetsByStatusQueryResponse, FindPetsByStatus400, unknown>({ method: 'GET', url: '/pet/findByStatus', params, ...config })
-  return res.data
+async function findPetsByStatus({ params }: {
+    params?: FindPetsByStatusQueryParams;
+}, config: Partial<RequestConfig> = {}) {
+    const res = await client<FindPetsByStatusQueryResponse, FindPetsByStatus400, unknown>({ method: "GET", url: `/pet/findByStatus`, params, ...config });
+    return res.data;
 }
 
-export function findPetsByStatusQueryOptions(
-  {
-    params,
-  }: {
-    params?: MaybeRef<FindPetsByStatusQueryParams>
-  },
-  config: Partial<RequestConfig> = {},
-) {
-  const queryKey = findPetsByStatusQueryKey(params)
-  return queryOptions({
-    queryKey,
-    queryFn: async ({ signal }) => {
-      config.signal = signal
-      return findPetsByStatus(unref({ params: unref(params) }), unref(config))
-    },
-  })
+ export function findPetsByStatusQueryOptions({ params }: {
+    params?: MaybeRef<FindPetsByStatusQueryParams>;
+}, config: Partial<RequestConfig> = {}) {
+    const queryKey = findPetsByStatusQueryKey(params);
+    return queryOptions({
+        queryKey,
+        queryFn: async ({ signal }) => {
+            config.signal = signal;
+            return findPetsByStatus(unref({ params: unref(params) }), unref(config));
+        },
+    });
 }
 
-/**
+ /**
  * @description Multiple status values can be provided with comma separated strings
  * @summary Finds Pets by status
  * {@link /pet/findByStatus}
  */
-export function useFindPetsByStatus<
-  TData = FindPetsByStatusQueryResponse,
-  TQueryData = FindPetsByStatusQueryResponse,
-  TQueryKey extends QueryKey = FindPetsByStatusQueryKey,
->(
-  {
-    params,
-  }: {
-    params?: MaybeRef<FindPetsByStatusQueryParams>
-  },
-  options: {
-    query?: Partial<QueryObserverOptions<FindPetsByStatusQueryResponse, FindPetsByStatus400, TData, TQueryData, TQueryKey>>
-    client?: Partial<RequestConfig>
-  } = {},
-) {
-  const { query: queryOptions, client: config = {} } = options ?? {}
-  const queryKey = queryOptions?.queryKey ?? findPetsByStatusQueryKey(params)
-  const query = useQuery({
-    ...(findPetsByStatusQueryOptions({ params }, config) as unknown as QueryObserverOptions),
-    queryKey: queryKey as QueryKey,
-    ...(queryOptions as unknown as Omit<QueryObserverOptions, 'queryKey'>),
-  }) as UseQueryReturnType<TData, FindPetsByStatus400> & {
-    queryKey: TQueryKey
-  }
-  query.queryKey = queryKey as TQueryKey
-  return query
+export function useFindPetsByStatus<TData = FindPetsByStatusQueryResponse, TQueryData = FindPetsByStatusQueryResponse, TQueryKey extends QueryKey = FindPetsByStatusQueryKey>({ params }: {
+    params?: MaybeRef<FindPetsByStatusQueryParams>;
+}, options: {
+    query?: Partial<QueryObserverOptions<FindPetsByStatusQueryResponse, FindPetsByStatus400, TData, TQueryData, TQueryKey>>;
+    client?: Partial<RequestConfig>;
+} = {}) {
+    const { query: queryOptions, client: config = {} } = options ?? {};
+    const queryKey = queryOptions?.queryKey ?? findPetsByStatusQueryKey(params);
+    const query = useQuery({
+        ...findPetsByStatusQueryOptions({ params }, config) as unknown as QueryObserverOptions,
+        queryKey: queryKey as QueryKey,
+        ...queryOptions as unknown as Omit<QueryObserverOptions, "queryKey">
+    }) as UseQueryReturnType<TData, FindPetsByStatus400> & {
+        queryKey: TQueryKey;
+    };
+    query.queryKey = queryKey as TQueryKey;
+    return query;
 }