+++ packages/plugin-msw/src/generators/mswGenerator.test.tsx
@@ -63,6 +63,7 @@ describe('mswGenerator operation', async () => {
         path: '.',
       },
       parser: 'data',
+      group: undefined,
       ...props.options,
     }
     const plugin = { options } as Plugin<PluginMsw>