+++ packages/plugin-svelte-query/src/types.ts
@@ -122,6 +122,7 @@ export type Options = {
 
 type ResolvedOptions = {
   output: Output
+  group: Options['group']
   client: Required<Omit<NonNullable<PluginReactQuery['options']['client']>, 'baseURL'>> & { baseURL?: string }
   parser: Required<NonNullable<Options['parser']>>
   paramsType: NonNullable<Options['paramsType']>