+++ packages/plugin-swr/src/generators/mutationGenerator.test.tsx
@@ -100,6 +100,7 @@ describe('mutationGenerator operation', async () => {
       output: {
         path: '.',
       },
+      group: undefined,
       ...props.options,
     }
     const plugin = { options } as Plugin<PluginSwr>