+++ packages/plugin-zod/src/generators/operationsGenerator.test.tsx
@@ -42,6 +42,7 @@ describe('operationsGenerator operations', async () => {
       output: {
         path: '.',
       },
+      group: undefined,
       ...props.options,
     }
     const plugin = { options } as Plugin<PluginZod>